elearning/requirements.txt
asgiref==3.8.1
channels==4.2.0
channels_redis==4.2.1
crispy-bootstrap5==2024.10
Daphne==4.0.0
Django==5.1.5
django-crispy-forms==2.3
django-extensions==3.2.3
djangorestframework==3.15.2
msgpack==1.1.0
pillow==11.1.0
redis==5.2.1
sqlparse==0.5.3
pytest>=7.0
pytest-django>=4.5
coverage>=6.0
whitenoise==6.4.0